Cleaned up with_items/Added restart after changes
Cleaned up unorganized with_items to be more clear
on what they are. Also added the ability to restart
mysql after changes are made and the cluster has
already been created. This currently works as
desired in different scenarios. However, there may
be a situation in which this may not work so this
may need to be adjusted later.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>