Allow the overriding of MariaDB's bind address.
Until now the role would automatically select the IP address of the
interface specified in the "galera_cluster_bind_interface" and use
both the galera cluster and MariaDB to bind to it.

This is fine for the galera part, but it results in MariaDB _not_
listening to 127.0.0.1. This was needed in my case because consul
connect only uses 127.0.0.1.

This change separates this behaviour by introducing the variable
"mariadb_bind_address". By default it takes the value of
"galera_cluster_bind_address" so no functionality changed. But you can
now override just "mariadb_bind_address" and set it to 0.0.0.0 for
example if you need normal MySQL clients to connect without interfering
with how galera operates.