Ensure specified NIC name exists in target system
If the variable "galera_cluster_bind_interface" is set to a
non-existent interface name, the role fails while creating the
galera "gcomm://" string with a cryptic "AnsibleUndefined is None"
error message. This assertion here ensures this never happens.

I spent about an hour troubleshooting why the gcomm string failed
during development so I hope with this nobody else needs to.