[debian] Stop constantly changing permissions of `/etc/mysql/conf.d/overrides.cnf`
For debian-based systems, `/etc/mysql/conf.d/` has its permissions
updated on every run to `0755`, and it recurses that, which will update
`/etc/mysql/conf.d/overrides.cnf` as well.  The very next task then
re-updates the permissions of that file to `0644`, which means every
single time the task runs, it reports as changed.

This change removes the `recurse: true` on the `ansible.builin.file`
task, which also brings the behavior inline with what happens for
redhat-based systems today.

Closes #150