Do not suppress change reporting of configuration directory
On both debian-based systems (`/etc/mysql/conf.d`) and redhat-based
systems (`/etc/my.cnf.d`), `changed_when` was set to `false` when
precreating the configuration directory before any overrides were added
via `mariadb_config_overrides`.

This change removes the `changed_when: false` lines for both systems, as
this is a change that should not be suppressed.  This issue hid a bug
that was reported in #150.